Sommelier: Further tidying
