--- src/main/scala/sommelier/App.scala
@@ -1,7 +1,0 @@
-package sommelier
-
-// todo more tests
-// todo restructure - DSL on sommelier package object, implementation stuff in packages
-// todo stress testing
-// todo events = dearboy, injection = beefish
-// todo why 405 instead of 404 on so many paths?
