+++ src/test/scala/sommelier/test/OutgoingMiddlewareTests.scala
@@ -1,8 +1,8 @@
 package sommelier.test
 
 import sommelier.routing.ApplyMiddleware.applyOutgoing
-import sommelier.{Method, Request}
-import sommelier.SyncResult.Accepted
+import sommelier.messaging.{Method, Request}
+import sommelier.routing.SyncResult.Accepted
 import sommelier.routing.Routing._
 import sommelier.test.support.CustomMatchers._
 import utest._
